requirements.txt
Flask
requests
PyPDF2
google-generativeai
gunicorn